requirements.txt
aiogram==3.3.0
SQLAlchemy==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
redis==5.0.4
qrcode==7.4.2
Pillow
pydub